Updates to 0.9.8 'stable'
FS#708: This modifies the last-minute fix that made it into r371 to allow 
logged in users to see projects that have others_view=1. The change 
inadvertently showed all projects names to all users, but this corrects that.

FS#572: Reports using a date range were not including events happening on 
the ending date. This corrects that, and also makes any missing range 
endpoints assume today.