the `get out of the (add|strip)slashes hell`
every superglobal is now striptslashed by default (from fix.inc.php, yeah I
meant this from the begining ...) if magic_quotes are *on*.

adodb already ALWAYS add slashes, so it's done twice (nonsense), and it's
SQL-safe.

it allows us to just remove any call to add/stripslashes in the WHOLE project.

though, we have to fix the databases, I've generated an upgrade script (it's
exhaustive : I've taken every textual record of the current database schema, and
generated the script with vim MACROS.)


this is extensively checked on my system, and works *perfectly*