lots of fixes, finish (almost) details.php templatization
the history part is *not* well templatized. it's a real pain, and I've done a
real nasty dirty thing (the old php is in the template, and that just sucks).
have to make it better ... but later.

the rest is clean, and I should have reduced database load drastically on that
page (especially by caching results for tpl_userlink, but not only).