FS#583, FS#372
I added some functions from Dokuwiki and a self-made function which removes all chars from a string that are not allowed (whitelist instead of blacklist).
The following chars are allowed: a-z A-Z . _ -, space, extended latin chars, Cyrillic and Greek chars.
If an invalid char is in the username, there is no error message but the username gets corrected. The corrected version is sent per mail.