FS#1179: Assign Users to Projects / hide projects
Now project permissions for anon users are separated in
* view project
* view tasks

Now a user might be able to see a project (for example to be able to anonymously add a new task), but not see any tasks (apart from those he opened, using the token-URL). If a user cannot see a project though, he also can't see any tasks in it.