FS#1190: LDAP integration
So how does it work? Configuration in admin area -> user registration.
If a user tries to login which does not exist yet but is stored on the LDAP server, the user will automatically created. If this user has 'userPassword' stored on LDAP too, it will be checked and the user only added if it is correct.
After user creation the permissions may be changed.