0.9.9: - Typo - fixed that the template class can use a non-existent theme - table cells for task fields now use nowrap
1.0:
FS#218. Use the upgrader, it will take a while. And don't ignore the backup hint this time. After running the upgrader, you have to set the lists for each field manually in the admin area -> fields.
This is only a first version, more field types will be added later. For now, please let's focus on *bug fixing*.