I hate to say this.. but "habemus captcha" :D
This is the captcha support that flyspray will have it uses the
Carnegie Mellon University (the very inventors of CAPTCHAS)
free webservice called "reCAPTCHA", it requires you to regsiter and get
a api key.

Q.A

a) why a third party service ? I dont trust $X institution/cooporation
they are evil/unstrustable !!

R: If you have an implemenation that

 - Is licensed under a non-viral license (BSD prefered)

 - your captcha stuff is solvabable by humans.

 - It cannot be decoded/broken by a 13 yo kid, nor is listed in already
   defeated captcha list.

 - it is based on science-proven research, not in some speudo random
   stuff you invented some day.

 - it is backend by some knowlegble institution.

 - it has audible captcha for visually impaired people ( remember that
   audio should'nt be easy to crack either :D )

- You are willing to mantain the code in flyspray as well to document
  it.

- You are willing to answer users questions about your code and take care
  of the bug reports in our BTS.

- Your implementation can "learn" how spammers attempts to bypass the
  captcha and adapt itself easily.

If and ***only if** you can provide such things, feel free to critize
it, otherwise dont use it. 

We at flyspray.org hate captchas very much, but sooner or later this
will be required, better do it now than later.